refactor(session): extract SessionEncryptionService from SessionManager (#4228 step 2)
- New src/services/session/encryption.ts: dedicated AES-256-GCM encryption service
- SessionPersistenceService delegates encryption to SessionEncryptionService
- SessionManager delegates encryptSecret/decryptSecret to persistence service
- Removed duplicated crypto code from both SessionManager and PersistenceService
- Zero behavior changes — pure delegation refactoring
- Gate: tsc ✅, build ✅, 5630 tests ✅
